+++ arcutils/cas/utils.py
@@ -5,8 +5,9 @@ from urllib.parse import urlencode, urljoin
 from django.core.urlresolvers import reverse
 
 from arcutils.response import get_redirect_location
-from arcutils.settings import get_setting
 from arcutils.types.option import Some, Null
+
+from .settings import get_setting
 
 
 CAS_NS_MAP = {
@@ -50,13 +51,13 @@ def parse_cas_tree(tree):
 
         <cas:serviceResponse xmlns:cas='http://www.yale.edu/tp/cas'>
             <cas:authenticationSuccess>
-                <cas:user>wbaldwin</cas:user>
+                <cas:user>bobs</cas:user>
                 <cas:attributes>
-                    <cas:uid>wbaldwin</cas:uid>
-                    <cas:display_name>Wyatt Baldwin</cas:display_name>
-                    <cas:given_name>Wyatt</cas:given_name>
-                    <cas:sn>Baldwin</cas:sn>
-                    <cas:mail>wbaldwin@pdx.edu</cas:mail>
+                    <cas:uid>bobs</cas:uid>
+                    <cas:display_name>Bob Smith</cas:display_name>
+                    <cas:given_name>Bob</cas:given_name>
+                    <cas:sn>Smith</cas:sn>
+                    <cas:mail>bobs@pdx.edu</cas:mail>
                 </cas:attributes>
             </cas:authenticationSuccess>
         </cas:serviceResponse>
@@ -79,7 +80,7 @@ def login_url(request):
         https://cas.host.example.com/login?service={service_url()}
 
     """
-    path = get_setting('CAS.login_path', 'login')
+    path = get_setting('login_path', 'login')
     params = {'service': service_url(request)}
     return make_cas_url(path, **params)
 
@@ -92,7 +93,7 @@ def logout_url(request):
         https://cas.host.example.com/logout
 
     """
-    path = get_setting('CAS.logout_path', 'logout')
+    path = get_setting('logout_path', 'logout')
     return make_cas_url(path)
 
 
@@ -103,7 +104,7 @@ def make_cas_url(*path, **params):
     path prefix of the base CAS URL will be overwritten.
 
     """
-    base_url = get_setting('CAS.base_url')
+    base_url = get_setting('base_url')
     path = posixpath.join(*path)
     url = urljoin(base_url, path)
     if params:
@@ -122,7 +123,7 @@ def redirect_url(request):
     otherwise, fall back to ``next`` parameter, referrer, or /.
 
     """
-    default = get_setting('CAS.redirect_url', default=None)
+    default = get_setting('redirect_url', default=None)
     url = get_redirect_location(request, default=default)
     return url