+++ arcutils/ldap/utils.py
@@ -1,22 +1,6 @@
 import ldap3
 from ldap3.utils.conv import escape_filter_chars as escape  # noqa
 
-from arcutils.settings import get_setting
-
-
-def get_ldap_settings(using='default'):
-    """Get LDAP settings for the specified connection.
-
-    Args:
-        using: The name of an LDAP connection specified in the project's
-            settings
-
-    Returns:
-        dict: Settings for connection
-
-    """
-    return get_setting('LDAP.{using}'.format(using=using))
-
 
 def setting_to_ldap3_attr(name):
     """Get named attribute from ldap3 pacakge.